--- .github/workflows/nvblox_ci.yml
@@ -1,10 +1,0 @@
-name: GPU Dummy Job
-on:
-  pull_request:
-jobs:
-  gpu-job:
-    name: GPU Job
-    runs-on: [self-hosted, gpu] # assumes you have a runner with the label "gpu"
-    steps:
-      - name: Run dummy command
-        run: sleep 10
