+++ .github/workflows/premerge.yml
@@ -1,0 +1,14 @@
+name: nvblox premerge
+on:
+  pull_request:
+jobs:
+  gpu-job:
+    name: Premerge
+    runs-on: [self-hosted, gpu] # GPU jobs will run on AWS
+    steps:
+      - name: cpp-unit-tests x86_64-Cuda12-Ubuntu22
+        run: |
+          ci/premerge.py --platform x86_64 --cuda-version 12 --ubuntu-version 22 --build-and-test cpp
+      - name: python-unit-tests x86_64-Cuda12-Ubuntu22
+        run: |
+          ci/premerge.py --platform x86_64 --cuda-version 12 --ubuntu-version 22 --build-and-test python
